# docs/swagger.yaml
basePath: /api
definitions:
  handlers.CreateCommentRequest:
    properties:
      content:
        example: Great sound!
        type: string
    type: object
...
  handlers.LoginRequest:
    properties:
      password:
        example: securepassword123
        type: string
      username:
        example: johndoe
        type: string
    type: object
  handlers.LogoutRequest:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: securepassword123
        type: string
      username:
        example: johndoe
        type: string
    type: object
  handlers.SetReactionRequest:
    properties:
      type:
        enum:
        - like
        - dislike
        example: like
        type: string
    type: object
  handlers.UpdateCommentRequest:
    properties:
      content:
        example: Updated comment content
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Audio sharing and streaming platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: SoundTube API
  version: "1.0"
paths:
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and return JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Invalid credentials
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User login
      tags:
      - authentication
  /api/auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidate user JWT token by adding to blacklist
      parameters:
      - description: JWT token to invalidate
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logout successful
        "400":
          description: Invalid input format
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - authentication
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account with username, email and password
      parameters:
      - description: Registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User successfully registered
        "400":
          description: Invalid input format or registration error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register a new user
      tags:
      - authentication
  /api/auth/verify-email:
    get:
      description: Verify user email address using verification token
      parameters:
      - description: Email verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Email verified successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Token is required or invalid
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Verify email
      tags:
      - authentication
  /api/comments/{id}:
    delete:
      description: Delete a comment by ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Comment deleted successfully
          schema:
            type: object
        "403":
          description: Forbidden - not comment owner
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Comment not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete comment
      tags:
      - comments
    patch:
      consumes:
      - application/json
      description: Update a comment by ID
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateCommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Comment updated successfully
          schema:
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden - not comment owner
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Comment not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update comment
      tags:
      - comments
...
  /api/sounds/{id}/comments:
    get:
      description: Get all comments for a specific sound
      parameters:
      - description: Sound ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of comments
          schema:
            items:
              type: object
            type: array
        "400":
          description: Invalid sound ID
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get sound comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Add a new comment to a sound
      parameters:
      - description: Sound ID
        in: path
        name: id
        required: true
        type: integer
      - description: Comment data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Comment created successfully
          schema:
            type: object
        "400":
          description: Invalid input or sound ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create comment
      tags:
      - comments
  /api/sounds/{id}/reactions:
    delete:
      description: Remove user's reaction from a sound
      parameters:
      - description: Sound ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Reaction deleted successfully
          schema:
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Reaction not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete sound reaction
      tags:
      - reactions
    get:
      description: Get all reactions for a specific sound
      parameters:
      - description: Sound ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of reactions
          schema:
            items:
              type: object
            type: array
        "400":
          description: Invalid sound ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get sound reactions
      tags:
      - reactions
    put:
      consumes:
      - application/json
      description: Set like or dislike reaction for a sound
      parameters:
      - description: Sound ID
        in: path
        name: id
        required: true
        type: integer
      - description: Reaction type
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SetReactionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Reaction set successfully
          schema:
            type: object
        "400":
          description: Invalid input or reaction type
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Set sound reaction
      tags:
      - reactions
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"